Add support for roid, classtar

+++ src/test/scala/com/astrolabsoftware/grafink/models/Alert.scala
@@ -27,7 +27,7 @@ case class Alert(
   rfscore: Double,
   snnscore: Double,
   classtar: Double,
-  roid: Double,
+  roid: Int,
   mulens_class_1: Option[String] = None,
   mulens_class_2: Option[String] = None,
   year: Int,
